requirements.txt
asgiref==3.5.2
backcall==0.2.0
certifi==2022.6.15.1
charset-normalizer==2.1.1
decorator==5.1.1
Django==4.1.6
django-admin-sortable2==2.1.4
django-proxy==1.2.1
idna==3.3
ipdb==0.13.9
ipython==7.34.0
jedi==0.18.1
logicore-django-react==1.0.0.dev17
logicore-django-react-pages==1.0.0.dev5
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.31
psycopg2==2.9.3
ptyprocess==0.7.0
Pygments==2.13.0
pytz==2022.2.1
requests==2.28.1
sqlparse==0.4.2
toml==0.10.2
traitlets==5.4.0
typing_extensions==4.3.0
urllib3==1.26.12
wcwidth==0.2.5
